+++ myems-admin/app/controllers/settings/space/spacephotovoltaicpowerstation.controller.js
@@ -3,6 +3,7 @@
 app.controller('SpacePhotovoltaicPowerStationController', function(
     $scope,
     $window,
+    $timeout,
     $translate,
     SpaceService,
     PhotovoltaicPowerStationService,
@@ -14,6 +15,7 @@ app.controller('SpacePhotovoltaicPowerStationController', function(
     $scope.spacephotovoltaicpowerstations = [];
     $scope.cur_user = JSON.parse($window.localStorage.getItem("myems_admin_ui_current_user"));
     $scope.isLoadingPhotovoltaicpowerstations = false;
+    $scope.tabInitialized = false;
 
     $scope.getAllSpaces = function() {
     let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
@@ -58,7 +60,7 @@ app.controller('SpacePhotovoltaicPowerStationController', function(
     SpacePhotovoltaicPowerStationService.getPhotovoltaicPowerStationsBySpaceID(id, headers, function (response) {
             $scope.isLoadingPhotovoltaicpowerstations = false;
             if (angular.isDefined(response.status) && response.status === 200) {
-              $scope.spacephotovoltaicpowerstations = $scope.spacephotovoltaicpowerstations.concat(response.data);
+              $scope.spacephotovoltaicpowerstations = response.data;
             } else {
               $scope.spacephotovoltaicpowerstations=[];
             }
@@ -127,8 +129,27 @@ app.controller('SpacePhotovoltaicPowerStationController', function(
 		});
 	};
 
-  $scope.getAllSpaces();
-	$scope.getAllPhotovoltaicPowerStations();
+    $scope.initTab = function() {
+        if (!$scope.tabInitialized) {
+            $scope.tabInitialized = true;
+            $scope.getAllSpaces();
+            $scope.getAllPhotovoltaicPowerStations();
+        }
+    };
+
+    $scope.$on('space.tabSelected', function(event, tabIndex) {
+        var TAB_INDEXES = ($scope.$parent && $scope.$parent.TAB_INDEXES) || { PHOTOVOLTAIC_POWER_STATION: 14 };
+        if (tabIndex === TAB_INDEXES.PHOTOVOLTAIC_POWER_STATION && !$scope.tabInitialized) {
+            $scope.initTab();
+        }
+    });
+
+    $timeout(function() {
+        var TAB_INDEXES = ($scope.$parent && $scope.$parent.TAB_INDEXES) || { PHOTOVOLTAIC_POWER_STATION: 14 };
+        if ($scope.$parent && $scope.$parent.activeTabIndex === TAB_INDEXES.PHOTOVOLTAIC_POWER_STATION && !$scope.tabInitialized) {
+            $scope.initTab();
+        }
+    }, 0);
 
   $scope.refreshSpaceTree = function() {
     let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };