+++ myems-admin/app/controllers/settings/space/spacepoint.controller.js
@@ -3,6 +3,7 @@
 app.controller('SpacePointController', function (
     $scope, 
     $window,
+    $timeout,
     $translate, 
     SpaceService,
     DataSourceService, 
@@ -17,6 +18,7 @@ app.controller('SpacePointController', function (
     $scope.datasources = [];
     $scope.points = [];
     $scope.isLoadingPoints = false;
+    $scope.tabInitialized = false;
 
     $scope.getAllSpaces = function() {
       let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
@@ -150,8 +152,27 @@ app.controller('SpacePointController', function (
         });
     };
 
-    $scope.getAllDataSources();
-    $scope.getAllSpaces();
+    $scope.initTab = function() {
+        if (!$scope.tabInitialized) {
+            $scope.tabInitialized = true;
+            $scope.getAllDataSources();
+            $scope.getAllSpaces();
+        }
+    };
+
+    $scope.$on('space.tabSelected', function(event, tabIndex) {
+        var TAB_INDEXES = ($scope.$parent && $scope.$parent.TAB_INDEXES) || { POINT: 4 };
+        if (tabIndex === TAB_INDEXES.POINT && !$scope.tabInitialized) {
+            $scope.initTab();
+        }
+    });
+
+    $timeout(function() {
+        var TAB_INDEXES = ($scope.$parent && $scope.$parent.TAB_INDEXES) || { POINT: 4 };
+        if ($scope.$parent && $scope.$parent.activeTabIndex === TAB_INDEXES.POINT && !$scope.tabInitialized) {
+            $scope.initTab();
+        }
+    }, 0);
 
     $scope.refreshSpaceTree = function() {
       let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };