+++ myems-admin/app/controllers/settings/space/spacesensor.controller.js
@@ -3,6 +3,7 @@
 app.controller('SpaceSensorController', function (
     $scope,
     $window,
+    $timeout,
     $translate,
     SpaceService,
     SensorService,
@@ -13,6 +14,7 @@ app.controller('SpaceSensorController', function (
     $scope.spacesensors = [];
     $scope.cur_user = JSON.parse($window.localStorage.getItem("myems_admin_ui_current_user"));
     $scope.isLoadingSensors = false;
+    $scope.tabInitialized = false;
 
     $scope.getAllSensors = function () {
         let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
@@ -125,8 +127,27 @@ app.controller('SpaceSensorController', function (
         });
     };
 
-    $scope.getAllSensors();
-    $scope.getAllSpaces();
+    $scope.initTab = function() {
+        if (!$scope.tabInitialized) {
+            $scope.tabInitialized = true;
+            $scope.getAllSensors();
+            $scope.getAllSpaces();
+        }
+    };
+
+    $scope.$on('space.tabSelected', function(event, tabIndex) {
+        var TAB_INDEXES = ($scope.$parent && $scope.$parent.TAB_INDEXES) || { SENSOR: 5 };
+        if (tabIndex === TAB_INDEXES.SENSOR && !$scope.tabInitialized) {
+            $scope.initTab();
+        }
+    });
+
+    $timeout(function() {
+        var TAB_INDEXES = ($scope.$parent && $scope.$parent.TAB_INDEXES) || { SENSOR: 5 };
+        if ($scope.$parent && $scope.$parent.activeTabIndex === TAB_INDEXES.SENSOR && !$scope.tabInitialized) {
+            $scope.initTab();
+        }
+    }, 0);
 
     $scope.refreshSpaceTree = function() {
       let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };